split test projects
