+++ services/web/server/tests/unit/with_dbs/01/test_guests_management.py
@@ -6,7 +6,6 @@ from copy import deepcopy
 
 import pytest
 from aiohttp import web
-
 from pytest_simcore.helpers.utils_projects import create_project
 from servicelib.application import create_safe_application
 from simcore_service_webserver import application
