+++ services/web/server/tests/unit/with_dbs/01/test_logout.py
@@ -1,5 +1,4 @@
 from aiohttp import web
-
 from pytest_simcore.helpers.utils_assert import assert_status
 from pytest_simcore.helpers.utils_login import LoggedUser
 from simcore_service_webserver.login.cfg import get_storage
