+++ services/web/server/tests/unit/with_dbs/09/test_projects.py
@@ -6,7 +6,6 @@ import json
 import time
 import unittest.mock as mock
 import uuid as uuidlib
-from asyncio import Future
 from copy import deepcopy
 from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 from unittest.mock import call
@@ -14,7 +13,7 @@ from unittest.mock import call
 import aiohttp
 import pytest
 import socketio
-from _helpers import ExpectedResponse, HTTPLocked, standard_role_response
+from _helpers import standard_role_response
 from aiohttp import web
 from aioresponses import aioresponses
 from models_library.projects_access import Owner
@@ -25,10 +24,8 @@ from models_library.projects_state import (
     RunningState,
 )
 from pytest_simcore.helpers.utils_assert import assert_status
-from pytest_simcore.helpers.utils_login import log_client_in
 from pytest_simcore.helpers.utils_mock import future_with_result
 from pytest_simcore.helpers.utils_projects import NewProject, delete_all_projects
-from servicelib import async_utils
 from servicelib.application import create_safe_application
 from simcore_service_webserver import catalog
 from simcore_service_webserver.db import setup_db
@@ -794,809 +791,3 @@ async def test_new_template_from_project(
                 uuidlib.UUID(node_name)
             except ValueError:
                 pytest.fail("Invalid uuid in workbench node {}".format(node_name))
-
-
-@pytest.mark.parametrize(*standard_role_response())
-@pytest.mark.parametrize(
-    "share_rights",
-    [
-        {"read": True, "write": True, "delete": True},
-        {"read": True, "write": True, "delete": False},
-        {"read": True, "write": False, "delete": False},
-        {"read": False, "write": False, "delete": False},
-    ],
-)
-async def test_share_project(
-    client,
-    logged_user: Dict,
-    primary_group: Dict[str, str],
-    standard_groups: List[Dict[str, str]],
-    all_group: Dict[str, str],
-    user_role: UserRole,
-    expected: ExpectedResponse,
-    storage_subsystem_mock,
-    mocked_director_subsystem,
-    catalog_subsystem_mock,
-    share_rights: Dict,
-    project_db_cleaner,
-):
-    # Use-case: the user shares some projects with a group
-
-    # create a few projects
-    new_project = await _new_project(
-        client,
-        expected.created,
-        logged_user,
-        primary_group,
-        project={"accessRights": {str(all_group["gid"]): share_rights}},
-    )
-    if new_project:
-        assert new_project["accessRights"] == {
-            str(primary_group["gid"]): {"read": True, "write": True, "delete": True},
-            str(all_group["gid"]): share_rights,
-        }
-
-        # user 1 can always get to his project
-        await _assert_get_same_project(client, new_project, expected.ok)
-
-    # get another user logged in now
-    user_2 = await log_client_in(
-        client, {"role": user_role.name}, enable_check=user_role != UserRole.ANONYMOUS
-    )
-    if new_project:
-        # user 2 can only get the project if user 2 has read access
-        await _assert_get_same_project(
-            client,
-            new_project,
-            expected.ok if share_rights["read"] else expected.forbidden,
-        )
-        # user 2 can only list projects if user 2 has read access
-        list_projects = await _list_projects(client, expected.ok)
-        assert len(list_projects) == (1 if share_rights["read"] else 0)
-        # user 2 can only update the project is user 2 has write access
-        project_update = deepcopy(new_project)
-        project_update["name"] = "my super name"
-        await _replace_project(
-            client,
-            project_update,
-            expected.ok if share_rights["write"] else expected.forbidden,
-        )
-        # user 2 can only delete projects if user 2 has delete access
-        await _delete_project(
-            client,
-            new_project,
-            expected.no_content if share_rights["delete"] else expected.forbidden,
-        )
-
-
-@pytest.fixture()
-def ensure_run_in_sequence_context_is_empty():
-    async_utils.sequential_jobs_contexts = {}
-
-
-# PUT --------
-@pytest.mark.parametrize(
-    "user_role,expected,expected_change_access",
-    [
-        (UserRole.ANONYMOUS, web.HTTPUnauthorized, web.HTTPUnauthorized),
-        (UserRole.GUEST, web.HTTPOk, web.HTTPForbidden),
-        (UserRole.USER, web.HTTPOk, web.HTTPOk),
-        (UserRole.TESTER, web.HTTPOk, web.HTTPOk),
-    ],
-)
-async def test_replace_project(
-    client,
-    logged_user,
-    user_project,
-    expected,
-    expected_change_access,
-    all_group,
-    ensure_run_in_sequence_context_is_empty,
-):
-    project_update = deepcopy(user_project)
-    project_update["description"] = "some updated from original project!!!"
-    await _replace_project(client, project_update, expected)
-
-    # replacing the owner access is not possible, it will keep the owner as well
-    project_update["accessRights"].update(
-        {str(all_group["gid"]): {"read": True, "write": True, "delete": True}}
-    )
-    await _replace_project(client, project_update, expected_change_access)
-
-
-@pytest.mark.parametrize(
-    "user_role,expected",
-    [
-        (UserRole.ANONYMOUS, web.HTTPUnauthorized),
-        (UserRole.GUEST, web.HTTPOk),
-        (UserRole.USER, web.HTTPOk),
-        (UserRole.TESTER, web.HTTPOk),
-    ],
-)
-async def test_replace_project_updated_inputs(
-    client, logged_user, user_project, expected, ensure_run_in_sequence_context_is_empty
-):
-    project_update = deepcopy(user_project)
-    #
-    # "inputAccess": {
-    #    "Na": "ReadAndWrite", <--------
-    #    "Kr": "ReadOnly",
-    #    "BCL": "ReadAndWrite",
-    #    "NBeats": "ReadOnly",
-    #    "Ligand": "Invisible",
-    #    "cAMKII": "Invisible"
-    #  },
-    project_update["workbench"]["5739e377-17f7-4f09-a6ad-62659fb7fdec"]["inputs"][
-        "Na"
-    ] = 55
-    await _replace_project(client, project_update, expected)
-
-
-@pytest.mark.parametrize(
-    "user_role,expected",
-    [
-        (UserRole.ANONYMOUS, web.HTTPUnauthorized),
-        (UserRole.GUEST, web.HTTPOk),
-        (UserRole.USER, web.HTTPOk),
-        (UserRole.TESTER, web.HTTPOk),
-    ],
-)
-async def test_replace_project_updated_readonly_inputs(
-    client, logged_user, user_project, expected, ensure_run_in_sequence_context_is_empty
-):
-    project_update = deepcopy(user_project)
-    project_update["workbench"]["5739e377-17f7-4f09-a6ad-62659fb7fdec"]["inputs"][
-        "Na"
-    ] = 55
-    project_update["workbench"]["5739e377-17f7-4f09-a6ad-62659fb7fdec"]["inputs"][
-        "Kr"
-    ] = 5
-    await _replace_project(client, project_update, expected)
-
-
-# DELETE -------
-
-
-@pytest.mark.parametrize(
-    "user_role,expected",
-    [
-        (UserRole.ANONYMOUS, web.HTTPUnauthorized),
-        (UserRole.GUEST, web.HTTPForbidden),
-        (UserRole.USER, web.HTTPNoContent),
-        (UserRole.TESTER, web.HTTPNoContent),
-    ],
-)
-async def test_delete_project(
-    client,
-    logged_user,
-    user_project,
-    expected,
-    storage_subsystem_mock,
-    mocked_director_subsystem,
-    catalog_subsystem_mock,
-    fake_services,
-):
-    # DELETE /v0/projects/{project_id}
-
-    fakes = fake_services(5)
-    mocked_director_subsystem[
-        "get_running_interactive_services"
-    ].return_value = future_with_result(fakes)
-
-    await _delete_project(client, user_project, expected)
-
-    if expected == web.HTTPNoContent:
-        mocked_director_subsystem[
-            "get_running_interactive_services"
-        ].assert_called_once()
-        calls = [call(client.server.app, service["service_uuid"]) for service in fakes]
-        mocked_director_subsystem["stop_service"].has_calls(calls)
-        # wait for the fire&forget to run
-        await asyncio.sleep(2)
-        await _assert_get_same_project(client, user_project, web.HTTPNotFound)
-
-
-@pytest.mark.parametrize(
-    "user_role,expected",
-    [
-        (UserRole.ANONYMOUS, web.HTTPUnauthorized),
-        (UserRole.GUEST, web.HTTPOk),
-        (UserRole.USER, web.HTTPOk),
-        (UserRole.TESTER, web.HTTPOk),
-    ],
-)
-async def test_open_project(
-    client,
-    logged_user,
-    user_project,
-    client_session_id,
-    expected,
-    mocked_director_subsystem,
-):
-    # POST /v0/projects/{project_id}:open
-    # open project
-
-    url = client.app.router["open_project"].url_for(project_id=user_project["uuid"])
-    resp = await client.post(url, json=client_session_id())
-    await assert_status(resp, expected)
-    if resp.status == web.HTTPOk.status_code:
-        dynamic_services = {
-            service_uuid: service
-            for service_uuid, service in user_project["workbench"].items()
-            if "/dynamic/" in service["key"]
-        }
-        calls = []
-        for service_uuid, service in dynamic_services.items():
-            calls.append(
-                call(
-                    client.server.app,
-                    project_id=user_project["uuid"],
-                    service_key=service["key"],
-                    service_uuid=service_uuid,
-                    service_version=service["version"],
-                    user_id=logged_user["id"],
-                )
-            )
-        mocked_director_subsystem["start_service"].assert_has_calls(calls)
-
-
-@pytest.mark.parametrize(*standard_role_response())
-async def test_close_project(
-    client,
-    logged_user,
-    user_project,
-    client_session_id,
-    expected,
-    mocked_director_subsystem,
-    fake_services,
-):
-    # POST /v0/projects/{project_id}:close
-    fakes = fake_services(5)
-    assert len(fakes) == 5
-    mocked_director_subsystem[
-        "get_running_interactive_services"
-    ].return_value = future_with_result(fakes)
-
-    # open project
-    client_id = client_session_id()
-    url = client.app.router["open_project"].url_for(project_id=user_project["uuid"])
-    resp = await client.post(url, json=client_id)
-
-    if resp.status == web.HTTPOk.status_code:
-        calls = [
-            call(client.server.app, user_project["uuid"], logged_user["id"]),
-        ]
-        mocked_director_subsystem["get_running_interactive_services"].has_calls(calls)
-        mocked_director_subsystem["get_running_interactive_services"].reset_mock()
-
-    # close project
-    url = client.app.router["close_project"].url_for(project_id=user_project["uuid"])
-    resp = await client.post(url, json=client_id)
-    await assert_status(resp, expected.no_content)
-    if resp.status == web.HTTPNoContent.status_code:
-        calls = [
-            call(client.server.app, user_project["uuid"], None),
-            call(client.server.app, user_project["uuid"], logged_user["id"]),
-        ]
-        mocked_director_subsystem["get_running_interactive_services"].has_calls(calls)
-        calls = [call(client.server.app, service["service_uuid"]) for service in fakes]
-        mocked_director_subsystem["stop_service"].has_calls(calls)
-
-
-@pytest.mark.parametrize(
-    "user_role, expected",
-    [
-        (UserRole.ANONYMOUS, web.HTTPUnauthorized),
-        (UserRole.GUEST, web.HTTPOk),
-        (UserRole.USER, web.HTTPOk),
-        (UserRole.TESTER, web.HTTPOk),
-    ],
-)
-async def test_get_active_project(
-    client,
-    logged_user,
-    user_project,
-    client_session_id,
-    expected,
-    socketio_client,
-    mocked_director_subsystem,
-):
-    # login with socket using client session id
-    client_id1 = client_session_id()
-    sio = None
-    try:
-        sio = await socketio_client(client_id1)
-        assert sio.sid
-    except SocketConnectionError:
-        if expected == web.HTTPOk:
-            pytest.fail("socket io connection should not fail")
-
-    # get active projects -> empty
-    get_active_projects_url = (
-        client.app.router["get_active_project"]
-        .url_for()
-        .with_query(client_session_id=client_id1)
-    )
-    resp = await client.get(get_active_projects_url)
-    data, error = await assert_status(resp, expected)
-    if resp.status == web.HTTPOk.status_code:
-        assert not data
-        assert not error
-
-    # open project
-    open_project_url = client.app.router["open_project"].url_for(
-        project_id=user_project["uuid"]
-    )
-    resp = await client.post(open_project_url, json=client_id1)
-    await assert_status(resp, expected)
-
-    resp = await client.get(get_active_projects_url)
-    data, error = await assert_status(resp, expected)
-    if resp.status == web.HTTPOk.status_code:
-        assert not error
-        assert ProjectState(**data.pop("state")).locked.value
-        assert data == user_project
-
-    # login with socket using client session id2
-    client_id2 = client_session_id()
-    try:
-        sio = await socketio_client(client_id2)
-        assert sio.sid
-    except SocketConnectionError:
-        if expected == web.HTTPOk:
-            pytest.fail("socket io connection should not fail")
-    # get active projects -> empty
-    get_active_projects_url = (
-        client.app.router["get_active_project"]
-        .url_for()
-        .with_query(client_session_id=client_id2)
-    )
-    resp = await client.get(get_active_projects_url)
-    data, error = await assert_status(resp, expected)
-    if resp.status == web.HTTPOk.status_code:
-        assert not data
-        assert not error
-
-
-@pytest.mark.parametrize(
-    "user_role, expected_ok, expected_forbidden",
-    [
-        (UserRole.ANONYMOUS, web.HTTPUnauthorized, web.HTTPUnauthorized),
-        (UserRole.GUEST, web.HTTPOk, web.HTTPForbidden),
-        (UserRole.USER, web.HTTPOk, web.HTTPForbidden),
-        (UserRole.TESTER, web.HTTPOk, web.HTTPForbidden),
-    ],
-)
-async def test_delete_multiple_opened_project_forbidden(
-    client,
-    logged_user,
-    user_project,
-    mocked_director_api,
-    mocked_dynamic_service,
-    socketio_client,
-    client_session_id,
-    user_role,
-    expected_ok,
-    expected_forbidden,
-    mocked_director_subsystem,
-):
-    # service in project = await mocked_dynamic_service(logged_user["id"], empty_user_project["uuid"])
-    service = await mocked_dynamic_service(logged_user["id"], user_project["uuid"])
-    # open project in tab1
-    client_session_id1 = client_session_id()
-    try:
-        sio1 = await socketio_client(client_session_id1)
-    except SocketConnectionError:
-        if user_role != UserRole.ANONYMOUS:
-            pytest.fail("socket io connection should not fail")
-    url = client.app.router["open_project"].url_for(project_id=user_project["uuid"])
-    resp = await client.post(url, json=client_session_id1)
-    await assert_status(resp, expected_ok)
-    # delete project in tab2
-    client_session_id2 = client_session_id()
-    try:
-        sio2 = await socketio_client(client_session_id2)
-    except SocketConnectionError:
-        if user_role != UserRole.ANONYMOUS:
-            pytest.fail("socket io connection should not fail")
-    await _delete_project(client, user_project, expected_forbidden)
-
-
-@pytest.mark.parametrize(
-    "user_role, create_exp, get_exp, deletion_exp",
-    [
-        (
-            UserRole.ANONYMOUS,
-            web.HTTPUnauthorized,
-            web.HTTPUnauthorized,
-            web.HTTPUnauthorized,
-        ),
-        (UserRole.GUEST, web.HTTPForbidden, web.HTTPOk, web.HTTPForbidden),
-        (UserRole.USER, web.HTTPCreated, web.HTTPOk, web.HTTPNoContent),
-        (UserRole.TESTER, web.HTTPCreated, web.HTTPOk, web.HTTPNoContent),
-    ],
-)
-async def test_project_node_lifetime(
-    client,
-    logged_user,
-    user_project,
-    create_exp,
-    get_exp,
-    deletion_exp,
-    mocked_director_subsystem,
-    storage_subsystem_mock,
-    mocker,
-):
-
-    mock_storage_api_delete_data_folders_of_project_node = mocker.patch(
-        "simcore_service_webserver.projects.projects_handlers.projects_api.delete_data_folders_of_project_node",
-        return_value=Future(),
-    )
-    mock_storage_api_delete_data_folders_of_project_node.return_value.set_result("")
-
-    # create a new dynamic node...
-    url = client.app.router["create_node"].url_for(project_id=user_project["uuid"])
-    body = {"service_key": "some/dynamic/key", "service_version": "1.3.4"}
-    resp = await client.post(url, json=body)
-    data, errors = await assert_status(resp, create_exp)
-    node_id = "wrong_node_id"
-    if resp.status == web.HTTPCreated.status_code:
-        mocked_director_subsystem["start_service"].assert_called_once()
-        assert "node_id" in data
-        node_id = data["node_id"]
-    else:
-        mocked_director_subsystem["start_service"].assert_not_called()
-    # create a new NOT dynamic node...
-    mocked_director_subsystem["start_service"].reset_mock()
-    url = client.app.router["create_node"].url_for(project_id=user_project["uuid"])
-    body = {"service_key": "some/notdynamic/key", "service_version": "1.3.4"}
-    resp = await client.post(url, json=body)
-    data, errors = await assert_status(resp, create_exp)
-    node_id_2 = "wrong_node_id"
-    if resp.status == web.HTTPCreated.status_code:
-        mocked_director_subsystem["start_service"].assert_not_called()
-        assert "node_id" in data
-        node_id_2 = data["node_id"]
-    else:
-        mocked_director_subsystem["start_service"].assert_not_called()
-
-    # get the node state
-    mocked_director_subsystem[
-        "get_running_interactive_services"
-    ].return_value = future_with_result(
-        [{"service_uuid": node_id, "service_state": "running"}]
-    )
-    url = client.app.router["get_node"].url_for(
-        project_id=user_project["uuid"], node_id=node_id
-    )
-    resp = await client.get(url)
-    data, errors = await assert_status(resp, get_exp)
-    if resp.status == web.HTTPOk.status_code:
-        assert "service_state" in data
-        assert data["service_state"] == "running"
-
-    # get the NOT dynamic node state
-    mocked_director_subsystem[
-        "get_running_interactive_services"
-    ].return_value = future_with_result("")
-
-    url = client.app.router["get_node"].url_for(
-        project_id=user_project["uuid"], node_id=node_id_2
-    )
-    resp = await client.get(url)
-    data, errors = await assert_status(resp, get_exp)
-    if resp.status == web.HTTPOk.status_code:
-        assert "service_state" in data
-        assert data["service_state"] == "idle"
-
-    # delete the node
-    mocked_director_subsystem[
-        "get_running_interactive_services"
-    ].return_value = future_with_result([{"service_uuid": node_id}])
-    url = client.app.router["delete_node"].url_for(
-        project_id=user_project["uuid"], node_id=node_id
-    )
-    resp = await client.delete(url)
-    data, errors = await assert_status(resp, deletion_exp)
-    if resp.status == web.HTTPNoContent.status_code:
-        mocked_director_subsystem["stop_service"].assert_called_once()
-        mock_storage_api_delete_data_folders_of_project_node.assert_called_once()
-    else:
-        mocked_director_subsystem["stop_service"].assert_not_called()
-        mock_storage_api_delete_data_folders_of_project_node.assert_not_called()
-
-    # delete the NOT dynamic node
-    mocked_director_subsystem["stop_service"].reset_mock()
-    mock_storage_api_delete_data_folders_of_project_node.reset_mock()
-    # mock_director_api_get_running_services.return_value.set_result([{"service_uuid": node_id}])
-    url = client.app.router["delete_node"].url_for(
-        project_id=user_project["uuid"], node_id=node_id_2
-    )
-    resp = await client.delete(url)
-    data, errors = await assert_status(resp, deletion_exp)
-    if resp.status == web.HTTPNoContent.status_code:
-        mocked_director_subsystem["stop_service"].assert_not_called()
-        mock_storage_api_delete_data_folders_of_project_node.assert_called_once()
-    else:
-        mocked_director_subsystem["stop_service"].assert_not_called()
-        mock_storage_api_delete_data_folders_of_project_node.assert_not_called()
-
-
-@pytest.mark.parametrize("user_role,expected", [(UserRole.USER, web.HTTPOk)])
-async def test_tags_to_studies(
-    client, logged_user, user_project, expected, test_tags_data, catalog_subsystem_mock
-):
-    catalog_subsystem_mock([user_project])
-    # Add test tags
-    tags = test_tags_data
-    added_tags = []
-    for tag in tags:
-        url = client.app.router["create_tag"].url_for()
-        resp = await client.post(url, json=tag)
-        added_tag, _ = await assert_status(resp, expected)
-        added_tags.append(added_tag)
-        # Add tag to study
-        url = client.app.router["add_tag"].url_for(
-            study_uuid=user_project.get("uuid"), tag_id=str(added_tag.get("id"))
-        )
-        resp = await client.put(url)
-        data, _ = await assert_status(resp, expected)
-        # Tag is included in response
-        assert added_tag.get("id") in data.get("tags")
-
-    # check the tags are in
-    user_project["tags"] = [tag["id"] for tag in added_tags]
-    data = await _assert_get_same_project(client, user_project, expected)
-
-    # Delete tag0
-    url = client.app.router["delete_tag"].url_for(tag_id=str(added_tags[0].get("id")))
-    resp = await client.delete(url)
-    await assert_status(resp, web.HTTPNoContent)
-    # Get project and check that tag is no longer there
-    user_project["tags"].remove(added_tags[0]["id"])
-    data = await _assert_get_same_project(client, user_project, expected)
-    assert added_tags[0].get("id") not in data.get("tags")
-
-    # Remove tag1 from project
-    url = client.app.router["remove_tag"].url_for(
-        study_uuid=user_project.get("uuid"), tag_id=str(added_tags[1].get("id"))
-    )
-    resp = await client.delete(url)
-    await assert_status(resp, expected)
-    # Get project and check that tag is no longer there
-    user_project["tags"].remove(added_tags[1]["id"])
-    data = await _assert_get_same_project(client, user_project, expected)
-    assert added_tags[1].get("id") not in data.get("tags")
-
-    # Delete tag1
-    url = client.app.router["delete_tag"].url_for(tag_id=str(added_tags[1].get("id")))
-    resp = await client.delete(url)
-    await assert_status(resp, web.HTTPNoContent)
-
-
-@pytest.mark.parametrize(*standard_role_response())
-async def test_open_shared_project_2_users_locked(
-    client,
-    logged_user: Dict,
-    shared_project: Dict,
-    socketio_client: Callable,
-    client_session_id: Callable,
-    user_role: UserRole,
-    expected: ExpectedResponse,
-    aiohttp_client,
-    mocker,
-    disable_gc_manual_guest_users,
-):
-    # Use-case: user 1 opens a shared project, user 2 tries to open it as well
-    mock_project_state_updated_handler = mocker.Mock()
-
-    client_1 = client
-    client_id1 = client_session_id()
-    client_2 = await aiohttp_client(client.app)
-    client_id2 = client_session_id()
-
-    # 1. user 1 opens project
-    sio_1 = await _connect_websocket(
-        socketio_client,
-        user_role != UserRole.ANONYMOUS,
-        client_1,
-        client_id1,
-        {SOCKET_IO_PROJECT_UPDATED_EVENT: mock_project_state_updated_handler},
-    )
-    expected_project_state = ProjectState(
-        locked={"value": False},
-        state=ProjectRunningState(value=RunningState.NOT_STARTED),
-    )
-    await _state_project(
-        client_1,
-        shared_project,
-        expected.ok if user_role != UserRole.GUEST else web.HTTPOk,
-        expected_project_state,
-    )
-    await _open_project(
-        client_1,
-        client_id1,
-        shared_project,
-        expected.ok if user_role != UserRole.GUEST else web.HTTPOk,
-    )
-    expected_project_state.locked.value = True
-    expected_project_state.locked.owner = Owner(
-        user_id=logged_user["id"],
-        first_name=(logged_user["name"].split(".") + [""])[0],
-        last_name=(logged_user["name"].split(".") + [""])[1],
-    )
-    # NOTE: there are 2 calls since we are part of the primary group and the all group
-    await _assert_project_state_updated(
-        mock_project_state_updated_handler,
-        shared_project,
-        expected_project_state,
-        0 if user_role == UserRole.ANONYMOUS else 2,
-    )
-
-    await _state_project(
-        client_1,
-        shared_project,
-        expected.ok if user_role != UserRole.GUEST else web.HTTPOk,
-        expected_project_state,
-    )
-
-    # 2. create a separate client now and log in user2, try to open the same shared project
-    user_2 = await log_client_in(
-        client_2, {"role": user_role.name}, enable_check=user_role != UserRole.ANONYMOUS
-    )
-    sio_2 = await _connect_websocket(
-        socketio_client,
-        user_role != UserRole.ANONYMOUS,
-        client_2,
-        client_id2,
-        {SOCKET_IO_PROJECT_UPDATED_EVENT: mock_project_state_updated_handler},
-    )
-    await _open_project(
-        client_2,
-        client_id2,
-        shared_project,
-        expected.locked if user_role != UserRole.GUEST else HTTPLocked,
-    )
-    await _state_project(
-        client_2,
-        shared_project,
-        expected.ok if user_role != UserRole.GUEST else web.HTTPOk,
-        expected_project_state,
-    )
-
-    # 3. user 1 closes the project
-    await _close_project(client_1, client_id1, shared_project, expected.no_content)
-    if not any(user_role == role for role in [UserRole.ANONYMOUS, UserRole.GUEST]):
-        # Guests cannot close projects
-        expected_project_state = ProjectState(
-            locked=ProjectLocked(value=False),
-            state=ProjectRunningState(value=RunningState.NOT_STARTED),
-        )
-
-    # we should receive an event that the project lock state changed
-    # NOTE: there are 3 calls since we are part of the primary group and the all group and user 2 is part of the all group
-    await _assert_project_state_updated(
-        mock_project_state_updated_handler,
-        shared_project,
-        expected_project_state,
-        0
-        if any(user_role == role for role in [UserRole.ANONYMOUS, UserRole.GUEST])
-        else 3,
-    )
-    await _state_project(
-        client_1,
-        shared_project,
-        expected.ok if user_role != UserRole.GUEST else web.HTTPOk,
-        expected_project_state,
-    )
-
-    # 4. user 2 now should be able to open the project
-    await _open_project(
-        client_2,
-        client_id2,
-        shared_project,
-        expected.ok if user_role != UserRole.GUEST else HTTPLocked,
-    )
-    if not any(user_role == role for role in [UserRole.ANONYMOUS, UserRole.GUEST]):
-        expected_project_state.locked.value = True
-        expected_project_state.locked.owner = Owner(
-            user_id=user_2["id"],
-            first_name=(user_2["name"].split(".") + [""])[0],
-            last_name=(user_2["name"].split(".") + [""])[1],
-        )
-    # NOTE: there are 3 calls since we are part of the primary group and the all group
-    await _assert_project_state_updated(
-        mock_project_state_updated_handler,
-        shared_project,
-        expected_project_state,
-        0
-        if any(user_role == role for role in [UserRole.ANONYMOUS, UserRole.GUEST])
-        else 3,
-    )
-    await _state_project(
-        client_1,
-        shared_project,
-        expected.ok if user_role != UserRole.GUEST else web.HTTPOk,
-        expected_project_state,
-    )
-
-
-@pytest.mark.parametrize(*standard_role_response())
-async def test_open_shared_project_at_same_time(
-    loop,
-    client,
-    logged_user: Dict,
-    shared_project: Dict,
-    socketio_client: Callable,
-    client_session_id: Callable,
-    user_role: UserRole,
-    expected: ExpectedResponse,
-    aiohttp_client,
-    disable_gc_manual_guest_users,
-):
-    NUMBER_OF_ADDITIONAL_CLIENTS = 20
-    # log client 1
-    client_1 = client
-    client_id1 = client_session_id()
-    sio_1 = await _connect_websocket(
-        socketio_client,
-        user_role != UserRole.ANONYMOUS,
-        client_1,
-        client_id1,
-    )
-    clients = [
-        {"client": client_1, "user": logged_user, "client_id": client_id1, "sio": sio_1}
-    ]
-    # create other clients
-    for i in range(NUMBER_OF_ADDITIONAL_CLIENTS):
-        client = await aiohttp_client(client.app)
-        user = await log_client_in(
-            client,
-            {"role": user_role.name},
-            enable_check=user_role != UserRole.ANONYMOUS,
-        )
-        client_id = client_session_id()
-        sio = await _connect_websocket(
-            socketio_client,
-            user_role != UserRole.ANONYMOUS,
-            client,
-            client_id,
-        )
-        clients.append(
-            {"client": client, "user": user, "client_id": client_id, "sio": sio}
-        )
-
-    # try opening projects at same time (more or less)
-    open_project_tasks = [
-        _open_project(
-            c["client"],
-            c["client_id"],
-            shared_project,
-            [
-                expected.ok if user_role != UserRole.GUEST else web.HTTPOk,
-                expected.locked if user_role != UserRole.GUEST else HTTPLocked,
-            ],
-        )
-        for c in clients
-    ]
-    results = await asyncio.gather(
-        *open_project_tasks,
-        return_exceptions=True,
-    )
-
-    # one should be opened, the other locked
-    if user_role != UserRole.ANONYMOUS:
-        num_assertions = 0
-        for data, error in results:
-            assert data or error
-            if error:
-                num_assertions += 1
-            elif data:
-                project_status = ProjectState(**data.pop("state"))
-                assert data == shared_project
-                assert project_status.locked.value
-                assert project_status.locked.owner.first_name in [
-                    c["user"]["name"] for c in clients
-                ]
-
-        assert num_assertions == NUMBER_OF_ADDITIONAL_CLIENTS
